day 15 complete (wip)

+++ AOC2022/Challenges/Challenge15/Point.cs
@@ -1,6 +1,6 @@
 ﻿namespace AOC2022.Challenges.Challenge15;
 
-public record Point(int X, int Y)
+public record Point(long X, long Y)
 {
     public static Point operator -(Point left, Point right) => new(left.X - right.X, left.Y - right.Y);
